requirements.txt
customtkinter
gspread
openai
oauth2client
pandas
python-dotenv
pyarrow